+++ src/components/modal-overlay/styles.module.css
@@ -4,7 +4,6 @@
     left: 0;
     width: 100%;
     height: 100%;
-    display: flex;
     justify-content: center;
     align-items: center;
     backdrop-filter:
