+++ src/index.css
@@ -7,3 +7,7 @@ li {
     max-width: 700px;
     padding: 0;
 }
+
+:burgermodal {
+    @extend: root;
+  }
